# Conflicts:
#	src/main/java/com/nsa/cubric/application/controllers/Admin.java
#	src/main/resources/sql/seed_images.sql

+++ src/main/java/com/nsa/cubric/application/domain/UserRating.java
@@ -3,7 +3,7 @@ package com.nsa.cubric.application.domain;
 public class UserRating {
 	private Long id;
 
-	private Integer userProfileId;
+	private Long userProfileId;
 
 	private Integer imageId;
 
@@ -12,7 +12,7 @@ public class UserRating {
 	public UserRating() {
 	}
 
-	public UserRating(Long id, Integer userProfileId, Integer imageId, Boolean response) {
+	public UserRating(Long id, Long userProfileId, Integer imageId, Boolean response) {
 		this.id = id;
 		this.userProfileId = userProfileId;
 		this.imageId = imageId;
@@ -27,11 +27,11 @@ public class UserRating {
 		this.id = id;
 	}
 
-	public Integer getUserProfileId() {
+	public Long getUserProfileId() {
 		return userProfileId;
 	}
 
-	public void setUserProfileId(Integer userProfileId) {
+	public void setUserProfileId(Long userProfileId) {
 		this.userProfileId = userProfileId;
 	}
